Renamed gql to api
