+++ backend/dataall/modules/notebooks/api/__init__.py
@@ -1,4 +1,4 @@
 """The package defines the schema for SageMaker notebooks"""
-from dataall.modules.notebooks.gql import input_types, mutations, queries, types, resolvers
+from dataall.modules.notebooks.api import input_types, mutations, queries, types, resolvers
 
 __all__ = ["types", "input_types", "queries", "mutations", "resolvers"]
