+++ backend/dataall/modules/notebooks/api/queries.py
@@ -1,6 +1,6 @@
 """The module defines GraphQL queries for the SageMaker notebooks"""
 from dataall.api import gql
-from dataall.modules.notebooks.gql.resolvers import (
+from dataall.modules.notebooks.api.resolvers import (
     get_notebook,
     list_notebooks,
     get_notebook_presigned_url
