+++ backend/dataall/modules/notebooks/api/types.py
@@ -1,6 +1,6 @@
 """Defines the object types of the SageMaker notebooks"""
 from dataall.api import gql
-from dataall.modules.notebooks.gql.resolvers import (
+from dataall.modules.notebooks.api.resolvers import (
     resolve_notebook_stack,
     resolve_notebook_status,
     resolve_user_role,
@@ -9,7 +9,7 @@ from dataall.modules.notebooks.gql.resolvers import (
 from dataall.api.Objects.Environment.resolvers import resolve_environment
 from dataall.api.Objects.Organization.resolvers import resolve_organization_by_env
 
-from dataall.modules.notebooks.gql.enums import SagemakerNotebookRole
+from dataall.modules.notebooks.api.enums import SagemakerNotebookRole
 
 SagemakerNotebook = gql.ObjectType(
     name="SagemakerNotebook",
